--- service/__init__.py
@@ -1,3 +1,0 @@
-'''Init the service module'''
-
-from service import game as GameService
