+++ services/__init__.py
@@ -1,0 +1,3 @@
+'''Init the service module'''
+
+from services import game as GameService
